# repos/vulhealth/schemas/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VulHealth API",
    "version": "1.0.0",
    "description": "Hospital management API — DELIBERATELY vulnerable for pentest training."
  },
  "servers": [{ "url": "http://localhost:3001" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "paths": {
...
    "/api/appointments": {
      "post": { "summary": "Book appointment", "security": [{"bearerAuth": []}] }
    },
    "/api/appointments/{id}": {
      "get": { "summary": "Appointment detail", "security": [{"bearerAuth": []}], "parameters": [{"name":"id","in":"path","required":true}] }
    },
    "/api/records/mine": { "get": { "summary": "My medical records", "security": [{"bearerAuth": []}] } },
    "/api/records/{id}": {
      "get": { "summary": "Get medical record by id", "security": [{"bearerAuth": []}], "parameters": [{"name":"id","in":"path","required":true}] }
    },
...
    "/api/admin/users": { "get": { "summary": "List all users", "security": [{"bearerAuth": []}] } },
...
    "/api/admin/logs": { "get": { "summary": "Read server logs", "security": [{"bearerAuth": []}] } },
...
  }
}